Move exo-player specific files from noplayer.drm to exoplayer.drm, and reduce scope where possible.
