+++ core/src/main/java/com/novoda/noplayer/exoplayer/drm/FrameworkDrmSession.java
@@ -1,7 +1,8 @@
-package com.novoda.noplayer.drm;
+package com.novoda.noplayer.exoplayer.drm;
 
 import com.google.android.exoplayer2.drm.DrmSession;
 import com.google.android.exoplayer2.drm.FrameworkMediaCrypto;
+import com.novoda.noplayer.drm.SessionId;
 
 interface FrameworkDrmSession extends DrmSession<FrameworkMediaCrypto> {
