+++ core/src/main/java/com/novoda/noplayer/player/PlayerFactory.java
@@ -6,8 +6,8 @@ import android.os.Looper;
 
 import com.novoda.noplayer.Player;
 import com.novoda.noplayer.drm.DrmHandler;
-import com.novoda.noplayer.drm.DrmSessionCreator;
-import com.novoda.noplayer.drm.DrmSessionCreatorFactory;
+import com.novoda.noplayer.exoplayer.drm.DrmSessionCreator;
+import com.novoda.noplayer.exoplayer.drm.DrmSessionCreatorFactory;
 import com.novoda.noplayer.drm.DrmType;
 import com.novoda.noplayer.exoplayer.NoPlayerExoPlayerCreator;
 import com.novoda.noplayer.mediaplayer.NoPlayerMediaPlayerCreator;
