+++ core/src/test/java/com/novoda/noplayer/player/PlayerFactoryTest.java
@@ -5,8 +5,8 @@ import android.content.Context;
 import com.novoda.noplayer.Player;
 import com.novoda.noplayer.drm.DownloadedModularDrm;
 import com.novoda.noplayer.drm.DrmHandler;
-import com.novoda.noplayer.drm.DrmSessionCreator;
-import com.novoda.noplayer.drm.DrmSessionCreatorFactory;
+import com.novoda.noplayer.exoplayer.drm.DrmSessionCreator;
+import com.novoda.noplayer.exoplayer.drm.DrmSessionCreatorFactory;
 import com.novoda.noplayer.drm.DrmType;
 import com.novoda.noplayer.drm.StreamingModularDrm;
 import com.novoda.noplayer.exoplayer.NoPlayerExoPlayerCreator;
